+++ plugins/infovis/src/amort/InfovisAmortizedRenderer.h
