+++ plugins/infovis/src/infovis.cpp
@@ -16,11 +16,11 @@
 
 // Use extra block for renderer, so clang format does not change include order. Eigen3 (used in MDSProjection)
 // does not compile when X11 header (used in SDFFont, which is used in renderers) is included before.
-#include "InfovisAmortizedRenderer.h"
 #include "ParallelCoordinatesRenderer2D.h"
 #include "ScatterplotMatrixRenderer2D.h"
 #include "TableHistogramRenderer2D.h"
 #include "TextureHistogramRenderer2D.h"
+#include "amort/InfovisAmortizedRenderer.h"
 
 namespace megamol::infovis {
 class InfovisPluginInstance : public megamol::core::utility::plugins::AbstractPluginInstance {
